Basic state for our CSCI4661 Demo Day.
