+++ src/screens/FoodDetailScreen.js
@@ -2,34 +2,33 @@ import React, { useContext } from 'react';
 import { StyleSheet, View, Text, Button, FlatList } from 'react-native'
 import { TouchableOpacity } from 'react-native-gesture-handler';
 import { Context } from "../contexts/FoodContext"
+import { UserContext } from "../contexts/UserFoodContext"
+// import { ShoppingContext } from '../contexts/shoppingContext';
 
 const HeroDetailScreen = (props) => {
 	
 const { levelHero, state } = useContext(Context);
+const { deleteHero, userState } = useContext(UserContext);
+
 
 const heroID = props.navigation.getParam("id");
 const message = props.navigation.getParam("message");
 
-const hero = state.find((hero) => {
+const hero = userState.find((hero) => {
 	return heroID === hero.id;
 
 
 })
-const levelupPrice = hero.level * 10;
-const updatedGold = hero.gold - levelupPrice;
-const updatedHealth = hero.maxHealth + Math.floor(Math.random() * 7) + 3
-const updatedLevel = hero.level + 1
-const updatedPower = hero.power + Math.floor(Math.random() * 5) + 1
 
 return <View> 
 	{message? <Text style={styles.message}> {message} </Text> : null}
 	<Text style={styles.name}>{hero.name}</Text>
-	<Text style={styles.stats}>level:{hero.level}</Text>
-	<Text style={styles.stats}>Health:{hero.currentHealth}/{hero.maxHealth}</Text>
-	<Text style={styles.stats}>power: {hero.power} </Text>
-	<Text style={styles.stats}>Gold is: {hero.gold}</Text>
+	<Text style={styles.stats}>Calories: {hero.calories}</Text>
+	<Text style={styles.stats}>Carbs: {hero.carbs}/{hero.protein}</Text>
+	<Text style={styles.stats}>Fat: {hero.fat} </Text>
+	<Text style={styles.stats}>Sugar: {hero.sugar}</Text>
 	<View style={styles.level}>
-	<Button title={'Level Up! (' + levelupPrice + ' Gold)'} onPress={() => {levelHero(hero.id, hero.name, updatedLevel, updatedPower, updatedHealth, updatedHealth, updatedGold)}} />
+	<Button title={'Delete item'} onPress={() => {props.navigation.navigate("Pantry"), deleteHero(hero.id)}} />
 	</View>
 </View>
