+++ src/screens/SignUpScreen.js
@@ -1,7 +1,7 @@
 import React, {useState} from 'react';
 import {View, StyleSheet} from 'react-native';
 import {Text, Input, Button} from 'react-native-elements';
-import Spacer from '../components/Spacer';
+import Spacer from '../../components/Spacer';
 import Constant from 'expo-constants';
 
 const SignUpScreen = (props) => {
